add FieldsSortable
