+++ lib/components/Form/Checkbox.js
@@ -129,7 +129,7 @@ var Checkbox = function (_Component) {
 
 Checkbox.propTypes = process.env.NODE_ENV !== "production" ? {
   name: _propTypes2.default.string.isRequired,
-  checked: _propTypes2.default.bool.isRequired,
+  checked: _propTypes2.default.bool,
   onChange: _propTypes2.default.func.isRequired,
   /** Omit label prop to render Checkbox without a label */
   label: _propTypes2.default.string
