+++ lib/components/Form/Fields.js
@@ -1,7 +1,7 @@
 'use strict';
 
 exports.__esModule = true;
-exports.renderField = undefined;
+exports.renderField = exports.RequiredCharacters = exports.RequiredText = undefined;
 
 var _extends = Object.assign || function (target) { for (var i = 1; i < arguments.length; i++) { var source = arguments[i]; for (var key in source) { if (Object.prototype.hasOwnProperty.call(source, key)) { target[key] = source[key]; } } } return target; };
 
@@ -73,7 +73,7 @@ function _taggedTemplateLiteralLoose(strings, raw) { strings.raw = raw; return s
  * Form fields
  */
 
-var RequiredText = _styledComponents2.default.span(_templateObject, function (_ref) {
+var RequiredText = exports.RequiredText = _styledComponents2.default.span(_templateObject, function (_ref) {
   var theme = _ref.theme;
   return theme.colors.primary.base;
 }, function (_ref2) {
@@ -81,7 +81,7 @@ var RequiredText = _styledComponents2.default.span(_templateObject, function (_r
   return theme.colors.primary.darker;
 });
 
-var RequiredCharacters = _styledComponents2.default.p(_templateObject2, function (_ref3) {
+var RequiredCharacters = exports.RequiredCharacters = _styledComponents2.default.p(_templateObject2, function (_ref3) {
   var theme = _ref3.theme;
   return theme.fonts.secondary;
 }, function (_ref4) {
@@ -338,7 +338,13 @@ var Fields = function Fields(_ref8) {
 
 Fields.displayName = 'fields';
 Fields.propTypes = process.env.NODE_ENV !== "production" ? {
+  /**
+   * callback to be called when a field updates fn(fieldName, event)
+   */
   onChange: _propTypes2.default.func,
+  /**
+   * fields config array
+   */
   fields: _propTypes2.default.array
 } : {};
