+++ lib/components/Form/TextareaMd/components/Toolbar/FormattingToolbar.js
@@ -97,7 +97,7 @@ var FormattingToolbar = function (_React$Component) {
 
       return _react2.default.createElement(
         _ToolbarButton2.default,
-        { onMouseDown: onMouseDown, active: isActive },
+        { onMouseDown: onMouseDown, active: isActive, key: type },
         _react2.default.createElement(
           Tooltip,
           { tooltip: buttonLabel, placement: 'top' },
@@ -115,7 +115,7 @@ var FormattingToolbar = function (_React$Component) {
 
       return _react2.default.createElement(
         _ToolbarButton2.default,
-        { onMouseDown: onMouseDown, active: isActive },
+        { onMouseDown: onMouseDown, active: isActive, key: type },
         _react2.default.createElement(
           Tooltip,
           { tooltip: tooltip, placement: 'top' },
