+++ package.json
@@ -16,6 +16,7 @@
     "color": "^3.1.2",
     "immutable": "^4.0.0-rc.12",
     "rc-checkbox": "^2.1.7",
+    "react-beautiful-dnd": "^11.0.5",
     "react-day-picker": "^7.3.2",
     "react-portal": "^4.2.0",
     "rehype-react": "^4.0.1",
