+++ src/components/Form/Readme.mdx
@@ -15,6 +15,7 @@ import Calendar from './Calendar';
 import Errors from './Errors';
 import getSelectStyling from './selectStyling';
 import Fields from './Fields';
+import FieldsSortable from './FieldsSortable';
 import { DateUtils } from 'react-day-picker';
 import Select from 'react-select';
 import { ThemeConsumer } from 'styled-components';
@@ -212,11 +213,22 @@ Provide a config in object format like this:
 fields = [
     {
       name: 'name',
-      type: 'text',
-      title: 'What is your name?',
-      label: 'Name',
-      required: true
-    },
+      type: 'group',
+      list: [
+        {
+          name: 'first_name',
+          type: 'text',
+          label: 'First Name',
+          renderWidth: 'full',
+        },
+        {
+          name: 'last_name',
+          type: 'time',
+          label: 'Last Name',
+          renderWidth: 'full',
+        },
+      ],
+    }
     {
       name: 'pitch',
       type: 'textarea',
@@ -249,10 +261,25 @@ fields = [
     const fieldsConfig = [
       {
         name: 'name',
-        type: 'text',
-        title: 'What is your name?',
-        label: 'Name',
-        required: true
+        type: 'group',
+        list: [
+          {
+            name: 'first_name',
+            type: 'text',
+            title: 'First Name',
+            label: 'First Name',
+            required: true,
+            renderWidth: 'full',
+          },
+          {
+            name: 'last_name',
+            type: 'text',
+            title: 'Last Name',
+            label: 'Last Name',
+            required: true,
+            renderWidth: 'full',
+          },
+        ],
       },
       {
         name: 'pitch',
@@ -313,6 +340,72 @@ fields = [
 
     return (
       <Fields fields={fieldsWithData} onChange={handleChange} />
+    );
+
+}}
+
+</Playground>
+
+## Sortable Drag&Drop Fields
+
+<Props of={FieldsSortable} />
+
+`FieldsSortable` behaves just like `Fields` with the only difference that it doesn't support grouped fields. It adds drag handles to each field and allows them to be ordered.
+
+<Playground>
+  {() => {
+    const [fieldsConfig, setFieldsConfig] = useState([
+      {
+        name: 'thing1',
+        type: 'text',
+        label: 'First thing',
+      },
+      {
+        name: 'thing2',
+        type: 'text',
+        label: 'Second thing',
+      },
+      {
+        name: 'thing3',
+        type: 'text',
+        label: 'Third thing',
+      },
+      {
+        name: 'thing4',
+        type: 'text',
+        label: 'Fourth thing',
+      },
+    ]);
+
+    const [formData, setFormData] = useState({});
+
+    const handleChange = (fieldName, event) => {
+      if (event && typeof event.preventDefault === 'function') {
+        event.preventDefault();
+      }
+
+      if (event && event.target && event.target.value) {
+        newFormData = {
+          ...formData,
+          [fieldName]: event.target.value || event.target.checked
+        }
+      }
+
+      setFormData(newFormData)
+    }
+
+    const handleSort = (newFieldsConfig) => {
+      console.log(newFieldsConfig)
+
+      setFieldsConfig(newFieldsConfig)
+    }
+
+    const fieldsWithData = fieldsConfig.map((field) => {
+          return { ...field, value: formData[field.name] };
+      });
+
+    return (
+      <FieldsSortable fields={fieldsWithData} onChange={handleChange} onSort={handleSort}/>
     );
 
 }}
